requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-heroku==0.3.1
django-phone-field==1.8.1
django-tastypie==0.14.3
django-test==0.4030
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-mimeparse==1.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.1.0
wrapt==1.12.1
